requirements.txt
flask==3.0.2
requests==2.31.0
pymongo==4.6.2
python-dotenv==1.0.1
dnspython==2.7.0
werkzeug>=3.0.0 